+++ .next/server/server-reference-manifest.js
@@ -1,1 +1,1 @@
-self.__RSC_SERVER_MANIFEST="{\n  \"node\": {},\n  \"edge\": {},\n  \"encryptionKey\": \"wzmFlXh1qk2GFx9hadVafVrz0Nf4rgWC/4PS1h0s/1U=\"\n}"
+self.__RSC_SERVER_MANIFEST="{\n  \"node\": {},\n  \"edge\": {},\n  \"encryptionKey\": \"pZA1jcLt0ZezCT3CMRELSMPi+l8o2bPEmc1of6Q2NmU=\"\n}"